Add PostgresLikeCommon
